requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
DateTime==4.3
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
nltk==3.6.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
tika==1.24
tqdm==4.62.2
urllib3==1.26.6
Werkzeug==2.0.1
zope.interface==5.4.0
